+++ src/Components/DeleteBeerDialog.js
